+++ src/dr_clients/include/utilities.h
@@ -17,11 +17,11 @@ extern bool nudge_instrument;
 bool filter_bb_level_from_list (module_t * head, instr_t * instr);
 bool filter_module_level_from_list (module_t * head, instr_t * instr);
 bool filter_range_from_list (module_t * head, instr_t * instr); /* can be used for function calls */
-bool filter_from_list(module_t * head, instr_t * instr, uint mode); /* can be used for clients who do not need to do extra processing after filter for each differently */
-bool filter_from_module_name(module_t * head, char * name, uint mode);
+bool filter_from_list(module_t * head, instr_t * instr, uint64 mode); /* can be used for clients who do not need to do extra processing after filter for each differently */
+bool filter_from_module_name(module_t * head, char * name, uint64 mode);
 
 /* other utility functions */
-bool get_offset_from_module(app_pc instr_addr, uint * offset);
-uint populate_conv_filename(char * dest,const char * folder,const char * name, const char * other_details);
+bool get_offset_from_module(app_pc instr_addr, uint64 * offset);
+uint64 populate_conv_filename(char * dest,const char * folder,const char * name, const char * other_details);
 
 #endif
